Measure allocations over a long period rather than a 100-tick window
